==> 2_meme_OCR.ipynb <==
!wget --header="Host: doc-0o-0g-drive-data-export.googleusercontent.com" --header="User-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.212 Safari/537.36" --header="Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9" --header="Accept-Language: en-US,en;q=0.9" --header="Cookie: AUTH_q9gllbu0nte95409rtp7q7s721mapf6v_nonce=vvim2copfnj4s" --header="Connection: keep-alive" "https://doc-0o-0g-drive-data-export.googleusercontent.com/download/b3n3f7ro91h5cr0rs39bilh5evv17sjb/768o5famgcgilpb6ac4qrt7scien22aj/1622371500000/5f250682-e6b5-4aab-827b-d896868e70dd/110819034128027560352/ADt3v-NxvzdHPn5Euk9c55oh3sWYpvKYmiz7B35xq5Vy82W1Z0HJ7U65UVPkDPhHQlCmiZcStHu-a5rbs01f59lF7sKTSTygXm6wdG57GTE8Y1AvemIZA2yQ9v0-k-shLui4w38Y2LxiK2b5KjEIfRKnes3HZ9UdxQQq8byqxQk52UAi7x_QR1gkn1Q_P1qHRyQ9-SDGf3GuCThgZFqApbJLm1EV4FAKw4YOZ4emfc454OMTsriV77CeKMCLw0jBLxFyNN42jR6nlUCbc1kdPaWKiyGV7L0WvJG674eiDORMMRabYJDkrhKJU4eGjCwd3WZqLu8nxG8Z?authuser=0&nonce=vvim2copfnj4s&user=110819034128027560352&hash=cqno48rsdfe6ilopvofn67rggi4vql4d" -c -O 'train-20210530T104616Z-001.zip'

# Import libraries

import numpy as np 
import pandas as pd
import easyocr
import os
import pickle
import shutil
import re
import warnings
from zipfile import ZipFile
warnings.filterwarnings("ignore")

import nltk 
nltk.download('words')
words = set(nltk.corpus.words.words())

nltk.download('stopwords')
stopwords = set(nltk.corpus.stopwords.words('english')) 
stopwords.remove('not')

#load model for english language
reader = easyocr.Reader(['en'])

# Read text from all images with OCR

def give_text_from_images(raw_path):# read text from images
    exception_data=[]
    dict_data={}                 # this dictionary contain image_name and text given by ocr
    root=os.listdir(raw_path)    #list all images in folder
    for image_name in root:
      path=raw_path+image_name
      try:
        output = reader.readtext(path)     # readtext method read text in images
        ocr_text=' '
        for i in output:      # output[0]=cordinates of text,output[1]=text ,output[2]=probability
          if i[2]>=0.25:       # select text which has probability greater than 0.25
            ocr_text+=i[1]+" "
        dict_data[image_name] =ocr_text
      except:                  # if images is not png, jpg in that case ignore it
        exception_data.append(image_name)
    return dict_data,exception_data    

# reading text from train folder images

# reading text from train folder
train_text,train_exception=give_text_from_images('/content/drive/MyDrive/Meme/data/train/')

# these images are in gif format  so I have removed these
train_exception

# save train_text in pickle file
pickle.dump(train_text, open("/content/drive/MyDrive/Meme/data/train_text", "wb"))
pickle.dump(train_exception, open("/content/drive/MyDrive/Meme/data/train_exception", "wb"))

# reading text from validation folder images

validation_text,validation_exception=give_text_from_images('/content/drive/MyDrive/Meme/data/validation/')

#these images are in gif format  so I have removed these
validation_exception

#save validation_text in pickle file
pickle.dump(validation_text, open("/content/drive/MyDrive/Meme/data/validation_text", "wb"))
pickle.dump(validation_exception, open("/content/drive/MyDrive/Meme/data/validation_exception", "wb"))

# reading text from test folder images

test_text,test_exception=give_text_from_images('/content/drive/MyDrive/Meme/data/test/')

#save test_text in pickle file
pickle.dump(test_text, open("/content/drive/MyDrive/Meme/data/test_text", "wb"))
pickle.dump(test_exception, open("/content/drive/MyDrive/Meme/data/test_exception", "wb"))

#these images are in gif format  so I have removed these
test_exception

# seperate dank and not dank meme

def fun(source,destination,data):
    # before this function create folder class_0 and class_1
    # for every class that we have in our  directory we check if it's dank_level=1 then we will move it in class_1
    # otherwise move in class_0
    # list_id for dank_level==1
    list_id=list(data[data.dank_level==1]['id'])
    if os.path.isdir(source):
        data_files = os.listdir(source)
        for file in data_files:
            file_temp=file.split(".")[0]
            if file_temp in list_id:
              #if dank_level=1 then we remove in class_1
              shutil.move(source+'/'+file,destination+'/class_1') 
            else:
              if file!=='class_1':
                shutil.move(source+'/'+file,destination+'/class_0')   
           

#reading csv file of data
data=pd.read_csv('/content/drive/MyDrive/Meme/data/data.csv')

#separating class 1 and 0
source_train = '/content/drive/MyDrive/Meme/data/train'
destination_train= '/content/drive/MyDrive/Meme/data/train'
# call function  to seperate class 1 and 0
fun(source_train,destination_train,data)

source_validation = '/content/drive/MyDrive/Meme/data/validation'
destination_validation='/content/drive/MyDrive/Meme/data/validation'
#call function to seperate class 1 and 0
fun(source_validation,destination_validation,data)

# read saved ocr text from train and validation images

train_text=pickle.load(open('/content/drive/MyDrive/Meme/data/train_text','rb'))
validation_text=pickle.load(open('/content/drive/MyDrive/Meme/data/validation_text','rb'))
test_text=pickle.load(open("/content/drive/MyDrive/Meme/data/test_text", "rb"))

train_text=pd.DataFrame(train_text.items(), columns=['id','ocr_text'])
validation_text=pd.DataFrame(validation_text.items(), columns=['id','ocr_text'])
test_text=pd.DataFrame(test_text.items(), columns=['id','ocr_text'])

# steps for ocr text preprocessing

1. convert lower case
2. correct some cases  like this Don\'t,dont,dontt
2. remove some puctuation
3. remove word starts with @ and mem , words length<1 and words which are not  in english dictionary and stopwords
4. remove all non-words 

#some text features are
1. length of characters in ocr_text and no of words in ocr_text
2.length of characters in ocr_preprocess_text and no of words in ocr_preprocess_text

def ocr_preprocess(text_data):
  for i,row in text_data.iterrows():
      value=row['ocr_text']
      value=value.strip()  #remove space at begining and ending
      text_data.loc[i,'ocr_text']=value   #from orignal text remove space
      value=value.lower()
      text_data.loc[i,'ocr_no_of_words']=len(value.split())    # count no_of_words before preprocessing 
      text_data.loc[i,'ocr_text_len']=len(value)                  #len of text
      
      # to handle  cases like this Don\'t,dont,dontt
      # replace Don\'t with do not use replace("n't", " not")
      # reference appliedai.com
      value=value.replace("won't", "will not").replace("cannot", "can not").replace("can't", "can not")\
                            .replace("n't", " not").replace("what's", "what is").replace("it's", "it is")\
                            .replace("'ve", " have").replace("i'm", "i am").replace("'re", " are")\
                            .replace("he's", "he is").replace("she's", "she is").replace("'ll", " will")\
                            .replace("'s", " is").replace("'ll", " will").replace("'m", " am")\
                            .replace("'t", " not").replace("'d", " would").replace("dont","do not ").replace("dont","do not")\
                            .replace("dontt","do not")
      # remove  \n,\t,(,),|,.,-,_,<,>,?,!,\,:,;,{,},*,#,%,
      value=re.sub(r'[\n\t()|.-_<>,?!\\:;{}*#%]+',' ', value)   
      # remove  word start with mem 
      # remove  word  start with @   like  @steve 
      # word less than len 1
      x=' '.join(x for x in value.split()  if   not x.startswith('@') and not x.startswith('mem') and  x in words and len(x)>1 and x not in stopwords )
      #remove puctuation from words  /,[,],_ like r/Minecraft,[BIDET,closeraoa_,ulone_slavic_boy, u/TheOriginalNav 
      #remove all digit
      x=re.sub("\W"," ",x)
      x=x.strip()
      text_data.loc[i,'ocr_preprocess']=x  
      text_data.loc[i,'ocr_preprocess_no_of_words']=len(x.split())
      text_data.loc[i,'ocr_preprocess_text_len']=len(x)
  return text_data


#call preprocess function
validation_text=ocr_preprocess(validation_text)
train_text=ocr_preprocess(train_text)
test_text=ocr_preprocess(test_text)

train_text.shape,validation_text.shape,test_text.shape

#seperate csv file for traing and validation

data=pd.read_csv('/content/drive/MyDrive/Meme/data/data.csv')
#drop duplicates from data
data=data.drop_duplicates(subset=['id'],keep='first').reset_index(drop=True)
data.shape

train_text['id']=train_text['id'].apply(lambda x:x.split('.')[0])
train_data=train_text.merge(data,on='id',how='inner')
train_data=train_data.drop_duplicates(subset=['id'],keep='first').reset_index(drop=True)


validation_text['id']=validation_text['id'].apply(lambda x:x.split('.')[0])
validation_data=validation_text.merge(data,on='id',how='inner')
validation_data=validation_data.drop_duplicates(subset=['id'],keep='first').reset_index(drop=True)


test_text['id']=test_text['id'].apply(lambda x:x.split('.')[0])
test_data=test_text.merge(data,on='id',how='inner')
test_data=test_data.drop_duplicates(subset=['id'],keep='first').reset_index(drop=True)

train_data.shape,validation_data.shape,test_data.shape

def title_preprocess(text_data):
  for i,row in text_data.iterrows():
      value=row['title']
      value=value.strip()  #remove space at begining and ending
      text_data.loc[i,'title']=value   #from orignal text remove space
      value=value.lower()
      text_data.loc[i,'title_no_of_words']=len(value.split())    # count no_of_words before preprocessing 
      text_data.loc[i,'title_len']=len(value)                  #len of text
      
      # to handle  cases like this Don\'t,dont,dontt
      # replace Don\'t with do not use replace("n't", " not")
      # reference appliedai.com
      value=value.replace("won't", "will not").replace("cannot", "can not").replace("can't", "can not")\
                            .replace("n't", " not").replace("what's", "what is").replace("it's", "it is")\
                            .replace("'ve", " have").replace("i'm", "i am").replace("'re", " are")\
                            .replace("he's", "he is").replace("she's", "she is").replace("'ll", " will")\
                            .replace("'s", " is").replace("'ll", " will").replace("'m", " am")\
                            .replace("'t", " not").replace("'d", " would").replace("dont","do not ").replace("dont","do not")\
                            .replace("dontt","do not").replace("me_irl","me irl").replace("Me_irl",'me irl')
      # remove  \n,\t,(,),|,.,-,_,<,>,?,!,\,:,;,{,},*,#,%,
      value=re.sub(r'[\n\t()|.-_<>,?!\\:;{}*#%]+',' ', value)   
      # remove  word start with mem 
      # remove  word  start with @   like  @steve 
      # word less than len 1
      x=' '.join(x for x in value.split()  if   not x.startswith('@') and not x.startswith('mem') and len(x)>1 )
      #remove puctuation from words  /,[,],_ like r/Minecraft,[BIDET,closeraoa_,ulone_slavic_boy, u/TheOriginalNav 
      #remove all digit
      x=re.sub("\W"," ",x)
      x=x.strip()
      text_data.loc[i,'title_preprocess']=x  
      text_data.loc[i,'title_preprocess_no_of_words']=len(x.split())
      text_data.loc[i,'title_preprocess_text_len']=len(x)
  return text_data


validation_data=title_preprocess(validation_data)
train_data=title_preprocess(train_data)
test_data=title_preprocess(test_data)

#add_path of each images in csv file
def add_path(temp,data):
    for i,row in data.iterrows():
      root='/content/'+temp+'/'   #/content/validation/
      if row['dank_level']==1:
        data.loc[i,'path']=root+'class_1/'+row['id']+'.jpg'
      elif row['dank_level']==0:
        data.loc[i,'path']=root+'class_0/'+row['id']+'.jpg'
      else: 
        data.loc[i,'path'] ='error'  
    return data
train_data=add_path('train',train_data)
validation_data=add_path('validation',validation_data)    

def test_data_path(data):
    for i,row in data.iterrows():
      root='/content/test/'
      data.loc[i,'path']=root+row['id']+'.jpg'
    return data
test_data=test_data_path(test_data)      


train_data['text']= train_data['title_preprocess'].astype(str)+" "+train_data['ocr_preprocess'].astype(str)
validation_data['text']=validation_data['title_preprocess'].astype(str)+" "+validation_data['ocr_preprocess'].astype(str)
test_data['text']=test_data['title_preprocess'].astype(str)+' '+test_data['ocr_preprocess'].astype(str)

train_data.head(2)

#save train_data and validation_data
train_data.to_csv('/content/drive/MyDrive/Meme/data/train_data.csv',index=False)
validation_data.to_csv('/content/drive/MyDrive/Meme/data/validation_data.csv',index=False)
test_data.to_csv('/content/drive/MyDrive/Meme/data/test_data.csv',index=False)

print(train_data.shape)
print(validation_data.shape)
print(test_data.shape)
