==> 3_meme_hsv.ipynb <==
import cv2
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import os
# references :https://realpython.com/python-opencv-color-spaces/

What Are Color Spaces?

In the most common color space, RGB (Red Green Blue), colors are represented in terms of their red, green, and blue components. In more technical terms, RGB describes a color as a tuple of three components. Each component can take a value between 0 and 255, where the tuple (0, 0, 0) represents black and (255, 255, 255) represents white.

HSV and HSL are descriptions of hue, saturation, and brightness/luminance, which are particularly useful for identifying contrast in images. These color spaces are frequently used in color selection tools in software and for web design.

how to simply segment an object from an image based on color in Python using OpenCV

# all the color space conversions OpenCV provides
flags = [i for i in dir(cv2) if i.startswith('COLOR_')]

print(flags[:20])
# The first characters after COLOR_ indicate the origin color space
# The characters after the 2 are the target color space

 HSV stands for Hue, Saturation, and Value (or brightness), and is a cylindrical color space. The colors, or hues, are modeled as an angular dimension rotating around a central, vertical axis, which represents the value channel. Values go from dark (0 at the bottom) to light at the top. The third axis, saturation, defines the shades of hue from least saturated, at the vertical axis, to most saturated furthest away from the center:

def feature_hsv(path,images):
    avg_red=[]
    avg_green=[]
    avg_blue=[]
    avg_h=[]
    avg_s=[]
    avg_v=[]
    id=[]
    exception_img=[]
    for i in images:
        # OpenCV by default reads images in BGR format you can use the cvtColor(image, flag) and the flag we looked at above to fix this:
      try:
        nemo_BGR = cv2.imread(path+i)
        nemo = cv2.cvtColor(nemo_BGR, cv2.COLOR_BGR2RGB)
        r, g, b = cv2.split(nemo)
        # HSV is a good choice of color space for segmenting by color
        hsv_nemo = cv2.cvtColor(nemo, cv2.COLOR_RGB2HSV)
        h, s, v = cv2.split(hsv_nemo)
        id.append(i)
        avg_red.append(r.mean())
        avg_green.append(g.mean())
        avg_blue.append(b.mean())
        avg_h.append(h.mean())
        avg_s.append(s.mean())
        avg_v.append(v.mean())
      except:
        exception_img.append(i)
    hsv_dict={'id':id,'avg_red':avg_red,'avg_green':avg_green,'avg_blue':avg_blue,
              'avg_h': avg_h,'avg_s': avg_s,'avg_v': avg_v} 
    hsv_data= pd.DataFrame(hsv_dict) 
    return hsv_data       

#extract hsv and RGB values form train data
images=os.listdir('/content/drive/MyDrive/Meme/data/training')
train_hsv_data=feature_hsv('/content/drive/MyDrive/Meme/data/training/',images)

#extract hsv and RGB values form validation data
validation_images=os.listdir('/content/drive/MyDrive/Meme/validation/')
validataion_hsv_data=feature_hsv('/content/drive/MyDrive/Meme/validation/',validation_images)

#save extracted features
train_hsv_data.to_csv('/content/drive/MyDrive/Meme/data/train_hsv_data.csv',index=False)
validataion_hsv_data.to_csv('/content/drive/MyDrive/Meme/data/validation_hsv_data.csv',index=False)
