==> 4_Meme_Tfidf_w2v.ipynb <==
!wget --header="Host: doc-0o-7c-docs.googleusercontent.com" --header="User-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/94.0.4606.81 Safari/537.36" --header="Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9" --header="Accept-Language: en-US,en;q=0.9" --header="Cookie: AUTH_gbsevnc13krov7rjgq2cj0s062jqolpo=06568700698331676125|1634615700000|klej5cj5jrhqi9vkvioj16q2bmde14in" --header="Connection: keep-alive" "https://doc-0o-7c-docs.googleusercontent.com/docs/securesc/ths62md7879q0rdb5g663rm4c96q2241/sr1phqn2ct71pa1m2jr75p8755ogd8g1/1634615700000/00484516897554883881/06568700698331676125/1pGd5tLwA30M7wkbJKdXHaae9tYVDICJ_?e=download&authuser=0" -c -O 'glove_vectors'

# Import Libraries

import scipy
import pandas as pd
import numpy as np
from numpy import zeros
import nltk
import pickle
from scipy.sparse import hstack
from sklearn.metrics import roc_auc_score
from scipy.stats import randint as sp_randint
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.feature_extraction.text import CountVectorizer
from tqdm import tqdm
from xgboost import XGBClassifier
from sklearn.model_selection import RandomizedSearchCV
from nltk.sentiment.vader import SentimentIntensityAnalyzer
from sklearn.metrics  import f1_score,roc_auc_score
import warnings
nltk.download('vader_lexicon')
warnings.filterwarnings("ignore")

#read train and validation data
train_data=pd.read_csv('/content/drive/MyDrive/Meme/data/train_data.csv')
validation_data=pd.read_csv("/content/drive/MyDrive/Meme/data/validation_data.csv")
test_data=pd.read_csv('/content/drive/MyDrive/Meme/data/test_data.csv')

#fill thumbnail.width missing values
mean_width=train_data['thumbnail.width'].mean()
train_data['thumbnail.width'].fillna(value=mean_width,inplace=True)
validation_data['thumbnail.width'].fillna(value=mean_width,inplace=True)
test_data['thumbnail.width'].fillna(value=mean_width,inplace=True)

# fill  thumbnail.height missing values
mean_height=train_data['thumbnail.height'].mean()
train_data['thumbnail.height'].fillna(value=mean_height,inplace=True)
validation_data['thumbnail.height'].fillna(value=mean_height,inplace=True)
test_data['thumbnail.height'].fillna(value=mean_height,inplace=True)


#calculate area of image
train_data['area']=train_data['thumbnail.height']*train_data['thumbnail.width']
validation_data['area']=validation_data['thumbnail.height']*validation_data['thumbnail.width']

#define sentiment scores on text
def sentiment_scores(data):
    sid = SentimentIntensityAnalyzer()
    neg=[]
    pos=[]
    neu=[]
    comp=[]
    for sentence in data['text'].values: 
        sentence_sentiment_score = sid.polarity_scores(sentence)
        comp.append(sentence_sentiment_score['compound'])
        neg.append(sentence_sentiment_score['neg'])
        pos.append(sentence_sentiment_score['pos'])
        neu.append(sentence_sentiment_score['neu'])
    return comp,neg,pos,neu

# Train data  sentiment_scores
train_comp,train_neg,train_pos,train_neu=sentiment_scores(train_data)
train_data['com']=train_comp
train_data['neg']=train_neg
train_data['pos']=train_pos
train_data['neu']=train_neu

#validation data sentiment_scores
validation_comp,validation_neg,validation_pos,validation_neu=sentiment_scores(validation_data)
validation_data['com']=validation_comp
validation_data['neg']=validation_neg
validation_data['pos']=validation_pos
validation_data['neu']=validation_neu

# test data sentiment_scores
test_comp,test_neg,test_pos,test_neu=sentiment_scores(test_data)
test_data['com']=test_comp
test_data['neg']=test_neg
test_data['pos']=test_pos
test_data['neu']=test_neu

Read Image features file

train_image_features=pd.read_csv('/content/drive/MyDrive/Meme/data/train_hsv_data.csv')
train_image_features['id']=train_image_features['id'].apply(lambda x:x.split(".")[0])
train_image_features.head(4)

validation_image_features=pd.read_csv('/content/drive/MyDrive/Meme/data/validation_hsv_data.csv')
validation_image_features['id']=validation_image_features['id'].apply(lambda x:x.split(".")[0])
validation_image_features.head(4)

Merge Image Features

train_data=train_data.merge(train_image_features,on='id',how='left')
print(train_data.shape)
train_data.head(3)

validation_data=validation_data.merge(validation_image_features,on='id',how='left')
print(validation_data.shape)
validation_data.head(3)

train_data['hour']=train_data['datetime_temp'].apply(lambda x:int(x[11:13]))
validation_data['hour']=validation_data['datetime_temp'].apply(lambda x:int(x[11:13]))
test_data['hour']=test_data['datetime_temp'].apply(lambda x:int(x[11:13]))

#numeric features for training
list_features=['ocr_no_of_words', 'ocr_text_len', 'ocr_preprocess_no_of_words', 'ocr_preprocess_text_len',
               'title_no_of_words', 'title_len','title_preprocess_no_of_words', 'title_preprocess_text_len',
              'thumbnail.height' , 'thumbnail.width','time_of_day','hour','area','com', 'neg', 'pos', 'neu',
               'avg_red','avg_green','avg_blue','avg_h','avg_s','avg_v']

def min_max_scaling(df,list_features,dict_mm={ }):
    
    result = df.copy()
    if not bool(dict_mm):  #if dict is empty then we are finding mean and max for train data
        for feature_name in list_features:
            max_value = df[feature_name].max()
            min_value = df[feature_name].min()
            dict_mm[feature_name]=[max_value,min_value]
            result[feature_name] = (df[feature_name] - min_value) / (max_value - min_value)
        return result,dict_mm   
    else:                    #use train data min and max
        for feature_name in list_features:
            max_value=dict_mm[feature_name][0]
            min_value=dict_mm[feature_name][1]
            result[feature_name] = (df[feature_name] - min_value) / (max_value - min_value)
        return result

# min-max scaling on numeric features
train_data,dict_mm=min_max_scaling(train_data,list_features)
validation_data=min_max_scaling(validation_data,list_features,dict_mm)

train_numeric=train_data[list_features]
validation_numeric=validation_data[list_features]


vectorizer1=TfidfVectorizer(min_df=2,ngram_range=(1,4))
vectorizer1.fit(train_data['text'].values)

train_data_text_tfidf = vectorizer1.transform(train_data['text'].values)
validation_data_text_tfidf = vectorizer1.transform(validation_data['text'].values)
test_data_text_tfidf  = vectorizer1.transform(test_data['text'].values)

print("After vectorizations")
print(train_data_text_tfidf.shape)
print(validation_data_text_tfidf.shape)
print(test_data_text_tfidf.shape)
print("="*100)


dictionary = dict(zip(vectorizer1.get_feature_names(), list(vectorizer1.idf_)))
tfidf_words = set(vectorizer1.get_feature_names())

with open('glove_vectors', 'rb') as f:
    model = pickle.load(f)
    glove_words =  set(model.keys())

# reference appliedai
# average Word2Vec
# compute average word2vec for each review.
tfidf_w2v_vectors_train = []; # the avg-w2v for each sentence/review is stored in this list
for sentence in tqdm(train_data['text'].values): # for each review/sentence
    vector = np.zeros(300) # as word vectors are of zero length
    tf_idf_weight =0; # num of words with a valid vector in the sentence/review
    for word in sentence.split(): # for each word in a review/sentence
        if (word in glove_words) and (word in tfidf_words):
            vec = model[word] # getting the vector for each word
            # here we are multiplying idf value(dictionary[word]) and the tf value((sentence.count(word)/len(sentence.split())))
            tf_idf = dictionary[word]*(sentence.count(word)/len(sentence.split())) # getting the tfidf value for each word
            vector += (vec * tf_idf) # calculating tfidf weighted w2v
            tf_idf_weight += tf_idf
    if tf_idf_weight != 0:
        vector /= tf_idf_weight
    tfidf_w2v_vectors_train.append(vector)

print(len(tfidf_w2v_vectors_train))


# average Word2Vec
# compute average word2vec for each review.
tfidf_w2v_vectors_validation = []; # the avg-w2v for each sentence/review is stored in this list
for sentence in tqdm(validation_data['text'].values): # for each review/sentence
    vector = np.zeros(300) # as word vectors are of zero length
    tf_idf_weight =0; # num of words with a valid vector in the sentence/review
    for word in sentence.split(): # for each word in a review/sentence
        if (word in glove_words) and (word in tfidf_words):
            vec = model[word] # getting the vector for each word
            # here we are multiplying idf value(dictionary[word]) and the tf value((sentence.count(word)/len(sentence.split())))
            tf_idf = dictionary[word]*(sentence.count(word)/len(sentence.split())) # getting the tfidf value for each word
            vector += (vec * tf_idf) # calculating tfidf weighted w2v
            tf_idf_weight += tf_idf
    if tf_idf_weight != 0:
        vector /= tf_idf_weight
    tfidf_w2v_vectors_validation.append(vector)

print(len(tfidf_w2v_vectors_validation))
len(tfidf_w2v_vectors_validation[0])

train_numeric=scipy.sparse.csr_matrix(train_numeric)
validation_numeric=scipy.sparse.csr_matrix(validation_numeric)

# merge two sparse matrices: https://stackoverflow.com/a/19710648/4084039
from scipy.sparse import hstack

X_train = hstack((tfidf_w2v_vectors_train,   train_numeric)).tocsr()
X_validation = hstack((tfidf_w2v_vectors_validation, validation_numeric)).tocsr()

print("Final Data matrix")
print(X_train.shape)
print(X_validation.shape)
print("="*100)

X_train = hstack((tfidf_w2v_vectors_train, train_numeric)).tocsr()
X_validation = hstack((tfidf_w2v_vectors_validation, validation_numeric)).tocsr()

print("Final Data matrix")
print(X_train.shape)
print(X_validation.shape)
print("="*100)

xgbc = XGBClassifier()
parameters = {'learning_rate':[0.0001, 0.001, 0.01, 0.1, 0.2, 0.3] ,
              'n_estimators':range(5,100),
              'colsample_bytree':[0.2,0.25,0.3]}
clf =  RandomizedSearchCV(xgbc, parameters, cv=3, scoring='roc_auc',return_train_score=True,n_jobs=-1)
clf.fit(X_train, train_data['dank_level'])


results = pd.DataFrame.from_dict(clf.cv_results_)
results[['mean_train_score','mean_test_score', 'param_learning_rate','param_n_estimators','param_colsample_bytree']]

Train xgboost on learning_rate=0.0001 ,n_estimator=74,colsample_bytree=0.3

xgbc_final = XGBClassifier( learning_rate=0.0001 ,n_estimator=74,colsample_bytree=0.3)
xgbc_final.fit(X_train, train_data['dank_level'])

train_prediction=xgbc_final.predict_proba(X_train)
print('train auc is=',roc_auc_score(np.array(train_data['dank_level']), train_prediction[:,1]))
validation_prediction=xgbc_final.predict_proba(X_validation)
print('validation auc is=',roc_auc_score(np.array(validation_data['dank_level']), validation_prediction[:,1]))
