==> 6_meme_bert.ipynb <==
!wget --header="Host: doc-10-7c-docs.googleusercontent.com" --header="User-Agent: Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/90.0.4430.212 Safari/537.36" --header="Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9" --header="Accept-Language: en-US,en;q=0.9" --header="Cookie: AUTH_gbsevnc13krov7rjgq2cj0s062jqolpo_nonce=vgdcsfp52t5m4" --header="Connection: keep-alive" "https://doc-10-7c-docs.googleusercontent.com/docs/securesc/ths62md7879q0rdb5g663rm4c96q2241/htppf4vboej968isgqfsq1vahttiph9q/1622633025000/00484516897554883881/06568700698331676125/13exfXiyiByluh1PfYK1EyZyizqxeCVG9?e=download&authuser=0&nonce=vgdcsfp52t5m4&user=06568700698331676125&hash=ui7jlmmmhg2cc1b50rc1cjsern614lfr" -c -O 'tokenization.py'

Import Libraries

import pandas as pd
import pickle
import tokenization
from tokenization import FullTokenizer # reference :appliedai

import tensorflow as tf
import tensorflow_hub as hub
from tensorflow.keras.layers import Input, Dense, Activation, Dropout
from tensorflow.keras.layers import  Flatten,Dropout,BatchNormalization,Conv1D, MaxPooling1D
from tensorflow.keras.models import Model
from tensorflow.keras.optimizers import Adam
from tensorflow.keras.losses import CategoricalCrossentropy,categorical_crossentropy
from tensorflow.keras.metrics import Accuracy
from tensorflow.keras import callbacks
from tensorflow.keras.models import Sequential
from tensorflow.keras import layers
from tensorflow.keras.layers import Embedding
from sklearn.metrics  import f1_score,roc_auc_score, auc

#read train and validation and test data
train_data=pd.read_csv('/content/drive/MyDrive/Meme/data/train_data.csv')
validation_data=pd.read_csv("/content/drive/MyDrive/Meme/data/validation_data.csv")
test_data=pd.read_csv('/content/drive/MyDrive/Meme/data/test_data.csv')

data=pd.read_csv('/content/drive/MyDrive/Meme/data/final_dank.csv')
data.shape

## Loading the Pretrained Model from tensorflow HUB
tf.keras.backend.clear_session()

max_seq_length = 40

#BERT takes 3 inputs

#this is input words. Sequence of words represented as integers
input_word_ids = tf.keras.layers.Input(shape=(max_seq_length,), dtype=tf.int32, name="input_word_ids")

#mask vector if you are padding anything
input_mask = tf.keras.layers.Input(shape=(max_seq_length,), dtype=tf.int32, name="input_mask")

#segment vectors. If you are giving only one sentence for the classification, total seg vector is 0. 
#If you are giving two sentenced with [sep] token separated, first seq segment vectors are zeros and 
#second seq segment vector are 1's
segment_ids = tf.keras.layers.Input(shape=(max_seq_length,), dtype=tf.int32, name="segment_ids")

#bert layer 
bert_layer = hub.KerasLayer("https://tfhub.dev/tensorflow/bert_en_uncased_L-12_H-768_A-12/1", trainable=False)
pooled_output, sequence_output = bert_layer([input_word_ids, input_mask, segment_ids])

bert_model = Model(inputs=[input_word_ids, input_mask, segment_ids], outputs=pooled_output)


bert_model.summary()

bert_model.output

#getting Vocab file
vocab_file = bert_layer.resolved_object.vocab_file.asset_path.numpy()
do_lower_case = bert_layer.resolved_object.do_lower_case.numpy()

#reference: applied ai
tokenizer=FullTokenizer(vocab_file,do_lower_case)

test_tokens=[]
train_tokens=[]
validation_tokens=[ ]
for i in train_data['text']:
  tokens=tokenizer.tokenize(i)
  train_tokens.append(tokens[0:38])
for j in test_data['text']:
  tokens=tokenizer.tokenize(i)
  test_tokens.append(tokens[0:38])
for k in validation_data['text']:
  tokens=tokenizer.tokenize(i)
  validation_tokens.append(tokens[0:38])

for index,i in enumerate(train_tokens):
    train_tokens[index]=np.array(['[CLS]',*i,'[SEP]'])
 
for index,i in enumerate(test_tokens):
    test_tokens[index]=np.array(['[CLS]',*i,'[SEP]'])  

for index,i in enumerate(test_tokens):
    test_tokens[index]=np.array(['[CLS]',*i,'[SEP]'])     

print(train_tokens[9])

def padding(data):
  for index,i in enumerate(data):
    sen_len=len(i)
    if sen_len<max_seq_length :
      diff=max_seq_length -sen_len
      data[index]=np.array([*i,*["[PAD]" for k in range(diff)]])
  return data    

train_tokens=padding(train_tokens)
validation_tokens=padding(validation_tokens)
test_tokens=padding(test_tokens)

def mask(data):
  mask_data=[]
  for i in data:
    mask_data.append(list(map(lambda x:0 if x=='[PAD]' else 1,i)))
  return mask_data


train_mask=mask(train_tokens)
test_mask =mask(test_tokens)
validation_mask=mask(validation_tokens)

train_segment=np.zeros((len(train_mask),max_seq_length ))
validation_segment=np.zeros((len(validation_mask),max_seq_length ))
test_segment=np.zeros((len(test_mask),max_seq_length ))
print(train_segment.shape,test_segment.shape,validation_segment.shape)

print(train_tokens[2:5])

def token_id(data):
  for index,i in enumerate(data):
    data[index]=np.array(tokenizer.convert_tokens_to_ids(i))
  return (data)

train_tokens=token_id(train_tokens)
test_tokens=token_id(test_tokens)
validation_tokens=token_id(validation_tokens)


train_mask=np.asarray(train_mask)
test_mask=np.asarray(test_mask)
validation_mask=np.asarray(validation_mask)

train_tokens=np.asarray(train_tokens)
test_tokens=np.asarray(test_tokens)
validation_tokens=np.asarray(validation_tokens)

# get the train output, BERT model will give one output so save in
# X_train_pooled_output
train_pooled_output=bert_model.predict([train_tokens,train_mask,train_segment])

# get the test output, BERT model will give one output so save in
# test_pooled_output
test_pooled_output=bert_model.predict([test_tokens,test_mask,test_segment])
validation_pooled_output=bert_model.predict([validation_tokens,validation_mask,validation_segment])

##save all your results to disk so that, no need to run all again. 
pickle.dump((train_pooled_output,validation_pooled_output, test_pooled_output),open('/content/drive/MyDrive/Meme/data/bert_output.pkl','wb'))

##create an NN and 
def give_model():
  model=Sequential()
  model.add(Conv1D(256, 3,strides=5, activation='relu',input_shape=(768,1)))
  model.add(Conv1D(128, 3,strides=4, activation='relu'))
  model.add(Conv1D(64, 4,strides=2, activation='relu'))
  model.add(Conv1D(32, 2, activation='relu'))
  model.add(Flatten())

  
  model.add(Dense(128,activation='relu',kernel_initializer='he_normal'))
  model.add(Dropout(0.6))
  model.add(BatchNormalization())
  model.add(Dense(64,activation='relu',kernel_initializer='he_normal'))
  model.add(Dropout(0.6))
  model.add(BatchNormalization())

  model.add(Dense(32,activation='relu',kernel_initializer='he_normal'))
  model.add(Dropout(0.5))
  
  model.add(Dense(2,activation='softmax'))
  model.compile(optimizer=Adam(learning_rate=0.0001),loss='categorical_crossentropy',metrics=['accuracy'])
  return model

model =give_model()
model.summary()

y_train_one=tensorflow.keras.utils.to_categorical(train_data['dank_level'],2)
y_validation_one=tensorflow.keras.utils.to_categorical(validation_data['dank_level'],2)
y_test_one=tensorflow.keras.utils.to_categorical(test_data['dank_level'],2)


train_pooled_output= train_pooled_output.reshape(train_pooled_output.shape[0], train_pooled_output.shape[1], 1)
validation_pooled_output= validation_pooled_output.reshape(validation_pooled_output.shape[0], validation_pooled_output.shape[1], 1)
test_pooled_output= test_pooled_output.reshape(test_pooled_output.shape[0], test_pooled_output.shape[1], 1)

print("Training shape---> ",train_pooled_output.shape)
print("Test shape    ---> ",test_pooled_output.shape)
print("validation shape    ---> ",validation_pooled_output.shape)


model.fit(train_pooled_output,
           y_train_one,batch_size=128,
           validation_data=(validation_pooled_output,y_validation_one),
           epochs=100)
